Back/mysite/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.1
Django==3.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
nest-asyncio==1.5.1
numpy==1.21.0
opencv-python==4.5.5.62
outcome==1.1.0
pandas==1.2.5
parso==0.8.2
pbr==5.6.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
pyaes==1.6.1
pycparser==2.20
Pygments==2.10.0
pykeyboard==0.1.5
pyOpenSSL==21.0.0
Pyrogram==1.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==302
pyzmq==22.3.0
requests==2.26.0
selenium==4.0.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
stevedore==3.3.0
TgCrypto==1.2.3
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
virtualenv==20.4.2
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==2.0.1
wsproto==1.0.0
